% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_building_download_url}
\alias{get_building_download_url}
\title{Get download URL for a state}
\usage{
get_building_download_url(state = unique(tigris::fips_codes$state), version)
}
\arguments{
\item{state}{USPS abbreviation}

\item{version}{1 or 2}
}
\value{
url
}
\description{
Get download URL for a state
}
