% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bldg_centroids.R
\name{get_buildings}
\alias{get_buildings}
\title{Download Building Footprints from Microsoft}
\usage{
get_buildings(
  state = unique(tigris::fips_codes$state),
  county_fips = NULL,
  centroid = FALSE,
  version = 2
)
}
\arguments{
\item{state}{USPS state abbreviation}

\item{county_fips}{vector of 3 digit county FIPS codes to filter data by (optional)}

\item{centroid}{return centroids of buildings (default FALSE)}

\item{version}{version of data (1 or 2, default 2)}
}
\value{
an sf data frame of building footprints, plus county FIP and name if county_fips is specified
}
\description{
Download Building Footprints from Microsoft
}
\examples{
\dontrun{bronx_centroids <- get_buildings(
  state = 'NY',
  county_fips = '005',
  centroid = TRUE
)}
}
