% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{st_read_remote}
\alias{st_read_remote}
\title{Use sf::st_read on a remote, compressed file}
\usage{
st_read_remote(url, compressed = TRUE, ...)
}
\arguments{
\item{url}{url of file}

\item{compressed}{is file compressed?}

\item{...}{params to be passed to sf::st_read}
}
\value{
output of sf::st_read
}
\description{
Use sf::st_read on a remote, compressed file
}
